+++ src/range_to_list.c
@@ -1,16 +1,16 @@
 #include <stdio.h>
 #include <stdlib.h>
 
-float *centered_rangelist(float list_center, float range, int size)
+double *centered_rangelist(double list_center, double range, int size)
 {
-	float *list = malloc(sizeof(float) * size);
+	double *list = malloc(sizeof(double) * size);
 	if (list == NULL)
 	{
 		return NULL;
 	}
 
-	float list_start = list_center - (range / 2);
-	for (float i = 0, n = range / size, k = n / 2; i < size; i++)
+	double list_start = list_center - (range / 2);
+	for (double i = 0, n = range / size, k = n / 2; i < size; i++)
 	{
 		list[(int)i] = list_start + k;
 		k += n;
@@ -18,22 +18,22 @@ float *centered_rangelist(float list_center, float range, int size)
 	return list;
 }
 
-float *length_rangelist(float range[2], int length)
+double *length_rangelist(double range[2], int length)
 {
 	if (range[0] >= range[1])
 	{
 		return NULL;
 	}
 
-	float *list = malloc(sizeof(float) * length);
+	double *list = malloc(sizeof(double) * length);
 	if (list == NULL)
 	{
 		return NULL;
 	}
 
-	float range_diff = range[1] - range[0];
-	float increase = range_diff / length;
-	float n = range[0];
+	double range_diff = range[1] - range[0];
+	double increase = range_diff / length;
+	double n = range[0];
 	for (int i = 0; i < length; i++)
 	{
 		list[i] = n;
